fix readme discrepencies, update and fix docker build process
