+++ env.example
@@ -25,9 +25,9 @@ HOST_PORT=8002
 # Nostr relay URLs for FROSTR protocol communication
 # Can be JSON array or comma-separated string
 # Examples:
-#   JSON: ["wss://relay.damus.io","wss://relay.primal.net"]
-#   CSV:  wss://relay.damus.io,wss://relay.primal.net
-RELAYS=["wss://relay.damus.io","wss://relay.primal.net","wss://relay.snort.social"]
+#   JSON: ["wss://relay.primal.net","wss://relay.damus.io"]
+#   CSV:  wss://relay.primal.net,wss://relay.damus.io
+RELAYS=["wss://relay.primal.net"]
 
 # =============================================================================
 # SECURITY CONFIGURATION
