+++ src/routes/recovery.ts
@@ -6,15 +6,19 @@ import {
   validateShare
 } from '@frostr/igloo-core';
 import { RouteContext } from './types.js';
+import { getSecureCorsHeaders } from './utils.js';
 
 export async function handleRecoveryRoute(req: Request, url: URL, context: RouteContext): Promise<Response | null> {
   if (!url.pathname.startsWith('/api/recover')) return null;
 
+  // Get secure CORS headers based on request origin
+  const corsHeaders = getSecureCorsHeaders(req);
+  
   const headers = {
     'Content-Type': 'application/json',
-    'Access-Control-Allow-Origin': '*',
+    ...corsHeaders,
     'Access-Control-Allow-Methods': 'GET, POST, PUT, DELETE, OPTIONS',
-    'Access-Control-Allow-Headers': 'Content-Type',
+    'Access-Control-Allow-Headers': 'Content-Type, Authorization, X-API-Key, X-Session-ID',
   };
 
   try {
